refactor : ModalTypes.ts 를 types.ts로 네이밍 변경
