+++ components/src/lib/Modal/ButtonSet.tsx
@@ -1,7 +1,7 @@
 import { DarkButton, LightButton } from "../common/Button";
 import { InputWrapper } from "../common/Input.style";
 import { ModalButtonGroup } from "./Modal.style";
-import { ModalProps } from "./ModalType";
+import { ModalProps } from "./types";
 
 export const ButtonSet = ({
   modalType,
