+++ components/src/lib/index.ts
@@ -1,4 +1,4 @@
 export { useModal } from "./Modal";
 export { Modal } from "./Modal/Modal";
-export type { ModalProps } from "./Modal/ModalType";
+export type { ModalProps } from "./Modal/types";
 export { ButtonSet } from "./Modal/ButtonSet";
